pyproject.toml
[tool.poetry]
name = "bom-common"
version = "0.1.0"
description = ""
authors = ["Christopher L. Crutchfield <ccrutchf@ucsd.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
bom-pipeline = {git = "https://github.com/UCSD-E4E/bom-pipeline", rev = "main"}
tqdm = "^4.66.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
